+++ static/css/styles.css
@@ -31,7 +31,7 @@ text-align: center;
     width: 500px;
     height: 600px;
     margin-left: 50px;
-    background-image: url("https://github.com/cbnuLeehyunwoo/Fine_Dust_Measurment/blob/Feature_Howact/static/img/Cheongju.png?raw=true");
+    background-image: url("https://github.com/cbnuLeehyunwoo/Fine_Dust_Measurment/blob/Feature_image/static/img/Cheongju.png?raw=true");
     background-repeat: no-repeat;
     background-size: contain;
     position: relative;
